+++ src/components/instant-apps-header/instant-apps-header.tsx
@@ -13,8 +13,11 @@
  */
 
 import { Component, Host, h, Prop, Element, State } from '@stencil/core';
-import { Event, EventEmitter, HostElement } from '@stencil/core/internal';
+import { Event, EventEmitter, HostElement, Watch } from '@stencil/core/internal';
 import { getElementDir } from '../../utils/languageUtil';
+import Sanitizer from '@esri/arcgis-html-sanitizer';
+
+import { widthBreakpoints } from '../../utils/breakpoints';
 
 const CSS = {
   base: 'instant-apps-header--standard',
@@ -34,6 +37,27 @@ const CSS = {
   shadow: true,
 })
 export class InstantAppsHeader {
+  private _sanitizer = new Sanitizer(
+    {
+      whiteList: {
+        h1: ['style'],
+        h2: ['style'],
+        h3: ['style'],
+        h4: ['style'],
+        h5: ['style'],
+        h6: ['style'],
+        img: ['style', 'src', 'width', 'height'],
+        pre: ['style'],
+        p: ['id', 'class', 'style'],
+        div: ['id', 'class', 'style', 'role'],
+        span: ['id', 'class', 'style', 'role'],
+        figure: ['class', 'style'],
+        header: ['id', 'class', 'style', 'role']
+      },
+    },
+    true,
+  );
+
   @Element()
   el: HostElement;
 
@@ -83,6 +107,7 @@ export class InstantAppsHeader {
    */
   @Prop({
     reflect: true,
+    mutable: true,
   })
   infoButton: boolean = false;
 
@@ -95,18 +120,23 @@ export class InstantAppsHeader {
   infoIsOpen: boolean = false;
 
   /**
-   * HTML code for custom headers. IMPORTANT: SANITIZE YOUR HTML BEFORE PASSING IT IN TO AVOID SECURITY VULNERABILITIES.
+   * HTML code for custom headers.
    */
   @Prop({
-    reflect: true,
+    mutable: true,
   })
   customHeaderHtml: string;
 
+  @Watch('customHeaderHtml')
+  sanitizeCustomHeaderHtml() {
+    this.customHeaderHtml = this._sanitizer.sanitize(this.customHeaderHtml);
+  }
+
   /**
-   * CSS styles to be used in conjuction with `custom-header-html`. IMPORTANT: SANITIZE YOUR CSS BEFORE PASSING IT IN TO AVOID SECURITY VULNERABILITIES.
+   * CSS styles to be used in conjuction with `custom-header-html`.
    */
   @Prop({
-    reflect: true,
+    mutable: true,
   })
   customHeaderCss: string;
 
@@ -119,12 +149,24 @@ export class InstantAppsHeader {
   fontFamily: string = 'var(--calcite-sans-family);';
 
   /**
+   * Mobile breakpoint value in pixels(px).
+   */
+  @Prop()
+  mobileWidthBreakpoint: number = widthBreakpoints.medium[1];
+
+  /**
    * Fires when the info button is clicked.
    */
   @Event({ cancelable: false }) infoIsOpenChanged: EventEmitter<boolean>;
 
+  @State()
+  initialScale: 's' | 'm' | 'l' = 'm';
+
   componentWillLoad() {
     this.dir = getElementDir(this.el);
+    this.handleMobileBreakpoints();
+    this.dir = getElementDir(this.el);
+    this.customHeaderHtml = this._sanitizer.sanitize(this.customHeaderHtml);
   }
 
   render() {
@@ -133,7 +175,7 @@ export class InstantAppsHeader {
     return (
       <Host>
         {this.customHeaderHtml ? (
-          [<style>{this.customHeaderCss}</style>, <div innerHTML={this.customHeaderHtml} />]
+          [<style>{this.customHeaderCss}</style>, <div id="customHeader" innerHTML={this.customHeaderHtml} />]
         ) : (
           <header
             class={`${CSS.base}${this.dir === 'rtl' ? ` ${CSS.flipRtl}` : ''}${this.logoImage && !hasEmptyLogo ? ` ${CSS.logoHeight}${this.logoScale}` : ` ${CSS.standardHeight}`}`}
@@ -172,5 +214,26 @@ export class InstantAppsHeader {
   toggleInfo(): void {
     this.infoIsOpen = !this.infoIsOpen;
     this.infoIsOpenChanged.emit(this.infoIsOpen);
+  }
+
+  mqlCallback(): (event: MediaQueryListEvent) => void {
+    return event => {
+      const { matches } = event;
+      if (matches) {
+        this.logoScale = 's';
+        return;
+      }
+      this.logoScale = this.initialScale;
+    };
+  }
+
+  handleMobileBreakpoints(): void {
+    this.initialScale = this.logoScale;
+    const mediaQuery = `(max-width: ${this.mobileWidthBreakpoint}px)`;
+    const mql = window.matchMedia(mediaQuery);
+    if (mql.matches) {
+      this.logoScale = 's';
+    }
+    mql.addEventListener('change', this.mqlCallback());
   }
 }