+++ src/interfaces/interfaces.d.ts
@@ -113,3 +113,11 @@ interface PieChartRampElement {
   infos: SymbolTableElementInfo[];
   preview?: HTMLElement;
 }
+
+export interface MobileWidthBreakpoint {
+  xsmall: number[];
+  small: number[];
+  medium: number[];
+  large: number[];
+  xlarge: number[];
+}